requirements.txt
# --- Core (Бот и API) ---
aiogram>=3.10.0         # Основная библиотека для бота (версия 3.x)
fastapi>=0.109.0        # Фреймворк для бэкенда WebApp
uvicorn[standard]>=0.27.0 # ASGI-сервер для запуска FastAPI

# --- Database (База данных) ---
sqlalchemy>=2.0.25      # ORM для работы с БД (асинхронная)
asyncpg>=0.29.0         # Асинхронный драйвер для PostgreSQL
alembic>=1.13.1         # Инструмент для миграций БД

# --- Config & Utils (Настройки и утилиты) ---
pydantic-settings>=2.1.0 # Удобное чтение .env файлов и валидация конфигов
python-dotenv>=1.0.1    # Загрузка переменных окружения

# --- HTTP Client (Для внешних запросов, если понадобятся) ---
httpx>=0.26.0           # Асинхронный HTTP-клиент (замена requests)